+++ src/ts/components/subtitleoverlay.ts
@@ -5,6 +5,7 @@ import {Label, LabelConfig} from './label';
 import {ComponentConfig, Component} from './component';
 import {ControlBar} from './controlbar';
 import { EventDispatcher } from '../eventdispatcher';
+import {DOM} from '../dom';
 
 /**
  * Overlays the player to display subtitles.
@@ -317,8 +318,9 @@ class ActiveSubtitleManager {
     let id = ActiveSubtitleManager.calculateId(event);
 
     let label = new SubtitleLabel({
-      // Prefer the HTML subtitle text if set, else use the plain text
-      text: event.html || event.text,
+      // Prefer the HTML subtitle text if set, else try generating a image tag as string from the image attribute,
+      // else use the plain text
+      text: event.html || ActiveSubtitleManager.generateImageTagText(event.image) || event.text,
     });
 
     // Create array for id if it does not exist
@@ -329,6 +331,18 @@ class ActiveSubtitleManager {
     this.activeSubtitleCueCount++;
 
     return label;
+  }
+
+  private static generateImageTagText(imageData: string): string {
+    if (!imageData) {
+      return;
+    }
+
+    const imgTag = new DOM('img', {
+      src: imageData,
+    });
+    imgTag.css('width', '100%');
+    return imgTag.get(0).outerHTML; // return the html as string
   }
 
   /**