+++ src/ts/player-events.d.ts
@@ -560,6 +560,7 @@ declare namespace bitmovin {
       end: number;
       text: string;
       html?: string;
+      image?: string;
       region?: string;
       regionStyle?: string;
       position?: {